# backend1/openapi.json
{"openapi":"3.1.0","info":{"title":"RAG Backend1 — Ingest & Retrieval","description":"Enterprise resume RAG (后端1): LangChain ingest → Qdrant + SQLAlchemy metadata, tenant-scoped retrieval. Pass `X-Tenant-Id` on every call (JWT owned by 后端2).","version":"0.1.0"},"paths":{"/api/v1/health":{"get":{"tags":["ops"],"summary":"Health","operationId":"health_api_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/api/v1/documents":{"get":{"tags":["documents"],"summary":"List Documents","operationId":"list_documents_api_v1_documents_get","parameters":[{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/IngestStatus"},{"type":"null"}],"title":"Status"}},{"name":"doc_type","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/DocType"},{"type":"null"}],"title":"Doc Type"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":200,"minimum":1,"default":50,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}},{"name":"X-Tenant-Id","in":"header","required":true,"schema":{"type":"string","description":"Tenant id (required)","title":"X-Tenant-Id"},"description":"Tenant id (required)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocumentListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/ingest":{"post":{"tags":["ingest"],"summary":"Create Ingest","operationId":"create_ingest_api_v1_ingest_post","parameters":[{"name":"X-Tenant-Id","in":"header","required":true,"schema":{"type":"string","description":"Tenant id (required)","title":"X-Tenant-Id"},"description":"Tenant id (required)"}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_create_ingest_api_v1_ingest_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IngestCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/ingest/{document_id}":{"get":{"tags":["ingest"],"summary":"Get Ingest Status","operationId":"get_ingest_status_api_v1_ingest__document_id__get","parameters":[{"name":"document_id","in":"path","required":true,"schema":{"type":"string","title":"Document Id"}},{"name":"X-Tenant-Id","in":"header","required":true,"schema":{"type":"string","description":"Tenant id (required)","title":"X-Tenant-Id"},"description":"Tenant id (required)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IngestStatusResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/retrieve":{"post":{"tags":["retrieve"],"summary":"Retrieve Endpoint","operationId":"retrieve_endpoint_api_v1_retrieve_post","parameters":[{"name":"X-Tenant-Id","in":"header","required":true,"schema":{"type":"string","description":"Tenant id (required)","title":"X-Tenant-Id"},"description":"Tenant id (required)"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RetrieveRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RetrieveResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_create_ingest_api_v1_ingest_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"},"doc_type":{"$ref":"#/components/schemas/DocType","default":"kb"},"sync":{"type":"boolean","title":"Sync","default":false}},"type":"object","required":["file"],"title":"Body_create_ingest_api_v1_ingest_post"},"DocType":{"type":"string","enum":["resume","jd","internship","kb","other"],"title":"DocType"},"DocumentListItem":{"properties":{"document_id":{"type":"string","title":"Document Id"},"tenant_id":{"type":"string","title":"Tenant Id"},"filename":{"type":"string","title":"Filename"},"doc_type":{"type":"string","title":"Doc Type"},"status":{"type":"string","title":"Status"},"chunk_count":{"type":"integer","title":"Chunk Count"},"error_message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Message"},"created_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"}},"type":"object","required":["document_id","tenant_id","filename","doc_type","status","chunk_count"],"title":"DocumentListItem"},"DocumentListResponse":{"properties":{"tenant_id":{"type":"string","title":"Tenant Id"},"total":{"type":"integer","title":"Total"},"items":{"items":{"$ref":"#/components/schemas/DocumentListItem"},"type":"array","title":"Items"}},"type":"object","required":["tenant_id","total","items"],"title":"DocumentListResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status"},"service":{"type":"string","title":"Service"},"env":{"type":"string","title":"Env"}},"type":"object","required":["status","service","env"],"title":"HealthResponse"},"IngestCreateResponse":{"properties":{"document_id":{"type":"string","title":"Document Id"},"status":{"type":"string","title":"Status"},"message":{"type":"string","title":"Message"}},"type":"object","required":["document_id","status","message"],"title":"IngestCreateResponse"},"IngestStatus":{"type":"string","enum":["pending","processing","succeeded","failed"],"title":"IngestStatus"},"IngestStatusResponse":{"properties":{"document_id":{"type":"string","title":"Document Id"},"tenant_id":{"type":"string","title":"Tenant Id"},"filename":{"type":"string","title":"Filename"},"doc_type":{"type":"string","title":"Doc Type"},"status":{"type":"string","title":"Status"},"chunk_count":{"type":"integer","title":"Chunk Count"},"error_message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Message"},"created_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"}},"type":"object","required":["document_id","tenant_id","filename","doc_type","status","chunk_count"],"title":"IngestStatusResponse"},"RetrieveHit":{"properties":{"score":{"type":"number","title":"Score"},"point_id":{"type":"string","title":"Point Id"},"text":{"type":"string","title":"Text"},"document_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Document Id"},"doc_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Doc Type"},"chunk_index":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Chunk Index"},"filename":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Filename"},"tenant_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Tenant Id"}},"type":"object","required":["score","point_id","text"],"title":"RetrieveHit"},"RetrieveRequest":{"properties":{"query":{"type":"string","maxLength":4000,"minLength":1,"title":"Query"},"top_k":{"type":"integer","maximum":50.0,"minimum":1.0,"title":"Top K","default":5},"doc_type":{"anyOf":[{"type":"string","enum":["resume","jd","internship","kb","other"]},{"type":"null"}],"title":"Doc Type"}},"type":"object","required":["query"],"title":"RetrieveRequest"},"RetrieveResponse":{"properties":{"tenant_id":{"type":"string","title":"Tenant Id"},"query":{"type":"string","title":"Query"},"hits":{"items":{"$ref":"#/components/schemas/RetrieveHit"},"type":"array","title":"Hits"}},"type":"object","required":["tenant_id","query","hits"],"title":"RetrieveResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}