# backend2/app/schemas/health.py
from pydantic import BaseModel


class DependencyStatus(BaseModel):
    name: str
    ok: bool
    detail: str | None = None


class HealthResponse(BaseModel):
    status: str
    service: str
    dependencies: list[DependencyStatus] = []